# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Cabin Reservation API
  version: 2.0.0
  description: |
    API completa para el sistema de reservas de cabañas de la cooperativa.
    
    ## Funcionalidades Principales
    
    - **Autenticación JWT**: Login, registro público, recuperación de contraseña
    - **Gestión de Reservas**: Creación, cancelación, confirmación con horarios
    - **Horarios de Check-in/Check-out**: Configurables por cabaña y reserva
    - **Gestión de Cabañas**: CRUD completo con horarios por defecto
    - **Disponibilidad**: Consulta de fechas disponibles con bloqueos obligatorios
    - **Lista de Espera**: Sistema de notificaciones y reclamo de prioridad
    - **Panel Administrativo**: Dashboard, métricas, gestión completa
    - **Precios Dinámicos**: Rangos de precios por temporada con historial
    
    ## Nuevas Funcionalidades v2.0
    
    - ✅ **Registro Público de Usuarios**: Nuevos usuarios pueden registrarse con documento válido
    - ✅ **Horarios de Check-in/Check-out**: Campos obligatorios en reservas y cabañas
    - ✅ **Validación de Documentos**: Verificación contra base de datos de asociados activos
    
  contact:
    name: Equipo de Desarrollo
    email: desarrollo@cooperativa.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo local
  - url: https://api.cooperativa.com
    description: Servidor de producción
security:
  - bearerAuth: []
paths:
  /api/auth/login:
    post:
      tags:
        - 🔐 Autenticación
      summary: Iniciar sesión
      description: Autentica un usuario existente y retorna tokens JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentLoginRequest'
            examples:
              admin:
                summary: Login de administrador
                value:
                  documentNumber: "12345678"
                  password: "password"
              professor:
                summary: Login de profesor
                value:
                  documentNumber: "87654321"
                  password: "password"
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/register:
    post:
      tags:
        - 🔐 Autenticación
      summary: Registro público de nuevo usuario
      description: |
        Registra un nuevo usuario validando su número de documento contra la base de datos de asociados activos.
        
        **Validaciones:**
        - Documento debe existir en la base de datos de asociados activos
        - Email debe ser único
        - Documento no puede estar ya registrado
        - PIN debe tener exactamente 4 dígitos
        - Rol debe ser PROFESSOR o RETIREE (no se permite ADMIN)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              new_user:
                summary: Registro de nuevo usuario
                value:
                  documentNumber: "99999999"
                  email: "nuevo.usuario@email.com"
                  name: "Nuevo Usuario"
                  phone: "+57-300-999-9999"
                  pin: "1234"
                  role: "PROFESSOR"
      responses:
        '200':
          description: Registro exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
              examples:
                success:
                  value:
                    accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    refreshToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '400':
          description: Solicitud inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_document:
                  summary: Documento no válido
                  value:
                    timestamp: "2024-01-01T10:00:00.000+00:00"
                    status: 400
                    error: "Bad Request"
                    message: "Número de documento no válido o deshabilitado"
                    path: "/api/auth/register"
                existing_document:
                  summary: Documento ya registrado
                  value:
                    timestamp: "2024-01-01T10:00:00.000+00:00"
                    status: 400
                    error: "Bad Request"
                    message: "Ya existe un usuario con este número de documento"
                    path: "/api/auth/register"
                existing_email:
                  summary: Email ya registrado
                  value:
                    timestamp: "2024-01-01T10:00:00.000+00:00"
                    status: 400
                    error: "Bad Request"
                    message: "Ya existe un usuario con este email"
                    path: "/api/auth/register"
  /api/auth/refresh:
    post:
      tags:
        - 🔐 Autenticación
      summary: Renovar token de acceso
      description: Genera un nuevo token de acceso usando el refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
                  description: Token de renovación
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Token renovado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
  /api/auth/recover-password:
    post:
      tags:
        - 🔐 Autenticación
      summary: Solicitar recuperación de contraseña
      description: Envía email con token para restablecer contraseña
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecoverPasswordRequest'
      responses:
        '200':
          description: Email de recuperación enviado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Email de recuperación enviado"
  /api/auth/reset-password:
    post:
      tags:
        - 🔐 Autenticación
      summary: Restablecer contraseña
      description: Restablece la contraseña usando el token de recuperación
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Contraseña restablecida exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Contraseña restablecida exitosamente"
  /api/auth/validate-token:
    post:
      tags:
        - 🔐 Autenticación
      summary: Validar token JWT
      description: Verifica si un token JWT es válido y no ha expirado
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateTokenRequest'
      responses:
        '200':
          description: Token válido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenValidationResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido del endpoint de login o registro
  schemas:
    DocumentLoginRequest:
      type: object
      required:
        - documentNumber
        - password
      properties:
        documentNumber:
          type: string
          description: Número de documento del usuario
          example: "12345678"
        password:
          type: string
          description: Contraseña del usuario
          example: "password"
    RegisterRequest:
      type: object
      required:
        - documentNumber
        - email
        - name
        - pin
      properties:
        documentNumber:
          type: string
          description: Número de documento válido en la base de datos de asociados
          minLength: 7
          maxLength: 20
          pattern: '^\d+$'
          example: "99999999"
        email:
          type: string
          format: email
          description: Email único del usuario
          maxLength: 255
          example: "nuevo.usuario@email.com"
        name:
          type: string
          description: Nombre completo del usuario
          minLength: 2
          maxLength: 100
          example: "Nuevo Usuario"
        phone:
          type: string
          description: Número de teléfono (opcional)
          maxLength: 20
          pattern: '^[+]?[\d\s\-()]+$'
          example: "+57-300-999-9999"
        pin:
          type: string
          description: PIN de 4 dígitos
          pattern: '^\d{4}$'
          example: "1234"
        role:
          type: string
          description: Rol del usuario
          enum: [PROFESSOR, RETIREE]
          example: "PROFESSOR"
    RecoverPasswordRequest:
      type: object
      required:
        - documentNumber
      properties:
        documentNumber:
          type: string
          description: Número de documento del usuario
          example: "12345678"
    ResetPasswordRequest:
      type: object
      required:
        - token
        - newPin
      properties:
        token:
          type: string
          description: Token de recuperación de contraseña
          example: "abc123def456"
        newPin:
          type: string
          description: Nuevo PIN de 4 dígitos
          pattern: '^\d{4}$'
          example: "5678"
    ValidateTokenRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: Token JWT a validar
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: Token de acceso JWT
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refreshToken:
          type: string
          description: Token de renovación JWT
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    TokenValidationResponse:
      type: object
      properties:
        valid:
          type: boolean
          description: Si el token es válido
          example: true
        expiresAt:
          type: string
          format: date-time
          description: Fecha y hora de expiración del token
          example: "2024-01-01T11:00:00.000Z"
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Fecha y hora del error
          example: "2024-01-01T10:00:00.000Z"
        status:
          type: integer
          description: Código de estado HTTP
          example: 400
        error:
          type: string
          description: Tipo de error
          example: "Bad Request"
        message:
          type: string
          description: Mensaje descriptivo del error
          example: "Solicitud inválida"
        path:
          type: string
          description: Ruta donde ocurrió el error
          example: "/api/reservations"